Support version naming (#33)
Support patching of the version displayName metadata property.

To update a version metadata, execute a PATCH request to
  /version/org/project/resource.html

with as data the 'url' where the resource can be downloaded and the
updated 'displayName'. For example the following payload:
{"url":"/versionsource/myorg/3246fe8b-3f32-481a-945a-786ee5a2133f/105ee238-c416-4371-896b-a16d1b0a4029.html","displayName": "My Named Version"}
The url used is the one from the version record as returned by the GET operation on /versionlist.